requirements.txt
Pillow
torch
torchvision
diffusers